PLA-245 Change file structure for husky hooks
Divide the types into separate files. Put the WithGitHooks interface at the top level (git folder).
