+++ src/common/git/husky/check-cargo-options.ts
@@ -1,0 +1,15 @@
+export interface CheckCargoOptions {
+  /**
+   * Perform a code formatting step
+   *
+   * @default true
+   */
+  readonly isFormatting?: boolean
+
+  /**
+   * Path to the cargo
+   *
+   * @default '.'
+   */
+  readonly workingDirectory?: string
+}
