+++ src/common/git/index.ts
@@ -1,2 +1,3 @@
 export * from './gitignore'
 export * from './husky'
+export * from './with-git-hooks'
